Flay

Description

The most commonly used names in this module are intended to be imported unqualified:

import Flay (Flay, Flayable(flay), Dict(Dict))

The rest of the names, qualified:

import qualified Flay

Documentation

type Flay c m s t f g = (forall a. Dict (c a) -> f a -> m (g a)) -> s -> m t

Flay c m s t f g allows converting s to t by replacing ocurrences of f with g by applicatively applying a function (forall a. c a => f a -> m (g a)) to targeted occurences of f a inside s.

A Flay must obey the inner and outer identity laws.

When using a Flay, m will be required to be a Functor in case the Flay targets one element, or an Applicative if it targets more than one. There will be no constraints on the rest of the arguments to Flay.

to flay: tr. v., to strip off the skin or surface of.

Example Flay:

data Foo f = Foo (f Int) (f Bool)

flayFoo :: (Applicative m, c Int, c Bool) => Flay c m (Foo f) (Foo g) f g
flayFoo h (Foo a b) = Foo <$> h Dict a <*> h Dict b

Example 1: Removing uncertaininy

Given the Foo above, consider the following values:

foo1 :: Foo Maybe
foo1 = Foo (Just 1) Nothing

foo2 :: Foo Maybe
foo2 = Foo (Just 2) (Just True)

It is possible to remove the uncertainty of the fields in Foo perhaps being empty (Nothing) by converting Foo Maybe to Foo Identity. However, we can't just write a function of type Foo Maybe -> Foo Identity because we have the possiblity of some of the fields being Nothing, like in foo1. Instead, we are looking for a function Foo Maybe -> Maybe (Foo Identity) which will result on Just only as long as all of the fields in Foo is Just, like in foo2. This is exactly what Applicative enables us to do:

fooMaybeToIdentity :: Foo Maybe -> Maybe (Foo Identity)
fooMaybeToIdentity (Foo a b) = Foo <$> fmap Identity a <*> fmap Identity b

In fact, notice that we are not really working with Just and Nothing directly, so we might as well leave Maybe polymorphic:

fooMToIdentity :: Applicative m => Foo m -> m (Foo Identity)
fooMToIdentity (Foo a b) = Foo <$> fmap Identity a <*> fmap Identity b

Flay, among other things, is intended to generalize this pattern so that what choice of Foo, Maybe or Identity you make, you can use Applicative this way. The easiest way to use Flay is through trivial', which is sufficient unless we need to enforce some constrain in the target elements wrapped in m inside foo (we don't need this now). With trivial', we could have defined fooFToIdentity this way:

fooMToIdentity :: Applicative m => Foo m -> m (Foo Identity)
fooMToIdentity = trivial' flayFoo (fmap Identity)

Some important things to notice here are that we are reusing flayFoo's knowledge of Foo's structure, and that the construction of the Identity applies to any value wrapped in m (Int and Bool in our case). Compare this last fact to traverse, where the target type must be the same.

Example 2: collect'

See the documentation for collect'. To sum up: for any given Flay, we can collect all of the Flay's targets into a Monoid, without knowing anything about the targets themselves beyond the fact that they satisfy a particular constraint.

Example 3: Reusable datatype shapes

We can use a Flay to repurpose a datatype while maintaining its "shape". For example, given our Foo example from before: Foo Identity represents the presence of two values Int and Char, Foo Maybe represents their potential presence, Foo [] represents the potential for zero or more Ints and Chars, while Foo (Const x) represent the presence of two values of type x. We can use flayFoo to convert between these representations. In all these cases, the shape of Foo is preserved, meaning we can continue to pattern match or project on it.

inner :: Flay Trivial Identity s s f f -> s -> s

Inner identity law:

(\fl -> runIdentity . trivial' fl pure) = id

outer :: Monad m => Flay Trivial m (m x) (m x) m m -> m x -> m x

Outer identity law:

(\fl -> join . trivial' fl pure) = id

Flayable

class Flayable c m s t f g | s -> f, t -> g, s g -> t, t f -> s where

Default Flay implementation for s and t.

When defining Flayable instances, you will usually leave c, m, f, and g polymomrphic.

flay must obey the inner and outer identity laws.

Methods

flay :: Flay c m s t f g

If s and g are instances of Generic, then flay gets a default implementation. For example, provided the Foo datatype shown in the documentation for Flay had a Generic instance, then the following Flayable instance would get a default implementation for flay:

instance (Applicative m, c Int, c Bool) => Flayable c m (Foo f) (Foo g) f g

Notice that while this default definition works for an s having "nested Flayables", GHC will prompt you for some additional constraints related to GFlay' in order for it to compile.

flay :: (Functor m, GFlay c m s t f g) => Flay c m s t f g

If s and g are instances of Generic, then flay gets a default implementation. For example, provided the Foo datatype shown in the documentation for Flay had a Generic instance, then the following Flayable instance would get a default implementation for flay:

instance (Applicative m, c Int, c Bool) => Flayable c m (Foo f) (Foo g) f g

Notice that while this default definition works for an s having "nested Flayables", GHC will prompt you for some additional constraints related to GFlay' in order for it to compile.

Generics

gflay :: (Functor m, GFlay c m s t f g) => Flay c m s t f g

class (Generic s, Generic t, GFlay' c m (Rep s) (Rep t) f g) => GFlay c m s t f g

Convenience Constraint for satisfying basic GFlay' needs for s and t.

class GFlay' c m s t f g where

Minimal complete definition

gflay'

Methods

gflay' :: Flay c m (s p) (t p) f g

Utils

class Trivial a

Constraint trivially satisfied by every type.

This can be used as the c parameter to Flay or Flayable in case you are not interested in observing the values inside f.

trivial

Arguments

:: Flayable Trivial m s t f g

=> (forall a. Trivial a => f a -> m (g a))

-> s

-> m t

Like trivial', but works on a Flayable instead of taking an explicit Flay.

trivial = trivial' flay

trivial'

Arguments

:: Flay Trivial m s t f g

-> (forall a. Trivial a => f a -> m (g a))

-> s

-> m t

You can use trivial' if you don't care about the c argument to Flay. This implies that you won't be able to observe the a in forall a. f a, all you can do with such a is pass it around.

trivial fl h
   = fl ((Dict :: Dict (Trivial a)) (fa :: f a) -> h fa)

collect

Arguments

:: (Monoid b, Flayable c (Const b) s t f (Const ()))

=> (forall a. Dict (c a) -> f a -> b)

-> s

-> b

Like collect', but works on a Flayable instead of an explicit Flay.

collect'

Arguments

:: Monoid b

=> Flay c (Const b) s t f (Const ())

-> (forall a. Dict (c a) -> f a -> b)

-> s

-> b

Collect all of the f a of the given Flay into a Monoid b.

Example usage, given Foo and flayFoo examples given in the documentation for Flay:

> collect' flayFoo
      (\(Dict :: Dict (Show a)) (Identity (a :: a)) -> [show a])
      (Foo (pure 4) (pure True))
["4","True"]

Re-exports

data Dict a :: Constraint -> * where

Values of type Dict p capture a dictionary for a constraint of type p.

e.g.

Dict :: Dict (Eq Int)

captures a dictionary that proves we have an:

instance Eq 'Int

Pattern matching on the Dict constructor will bring this instance into scope.

Constructors

Dict :: Dict a